+++ instrumentation/github.com/gin-gonic/gin/otelgin/option.go
@@ -27,21 +27,29 @@ type config struct {
 }
 
 // Option specifies instrumentation configuration options.
-type Option func(*config)
+type Option interface {
+	apply(*config)
+}
+
+type optionFunc func(*config)
+
+func (o optionFunc) apply(c *config) {
+	o(c)
+}
 
 // WithPropagators specifies propagators to use for extracting
 // information from the HTTP requests. If none are specified, global
 // ones will be used.
 func WithPropagators(propagators propagation.TextMapPropagator) Option {
-	return func(cfg *config) {
+	return optionFunc(func(cfg *config) {
 		cfg.Propagators = propagators
-	}
+	})
 }
 
 // WithTracerProvider specifies a tracer provider to use for creating a tracer.
 // If none is specified, the global provider is used.
 func WithTracerProvider(provider oteltrace.TracerProvider) Option {
-	return func(cfg *config) {
+	return optionFunc(func(cfg *config) {
 		cfg.TracerProvider = provider
-	}
+	})
 }