+++ instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo/mongo.go
@@ -17,13 +17,17 @@ package otelmongo
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"strings"
 	"sync"
 
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/codes"
+	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"go.mongodb.org/mongo-driver/bson"
+	"go.mongodb.org/mongo-driver/bson/bsontype"
 	"go.mongodb.org/mongo-driver/event"
 )
 
@@ -34,31 +38,36 @@ type spanKey struct {
 
 type monitor struct {
 	sync.Mutex
-	spans       map[spanKey]trace.Span
-	serviceName string
-	cfg         config
+	spans map[spanKey]trace.Span
+	cfg   config
 }
 
 func (m *monitor) Started(ctx context.Context, evt *event.CommandStartedEvent) {
+	var spanName string
+
 	hostname, port := peerInfo(evt)
 
 	attrs := []attribute.KeyValue{
-		ServiceName(m.serviceName),
-		DBOperation(evt.CommandName),
-		DBInstance(evt.DatabaseName),
-		DBSystem("mongodb"),
-		PeerHostname(hostname),
-		PeerPort(port),
+		semconv.DBSystemMongoDB,
+		semconv.DBOperationKey.String(evt.CommandName),
+		semconv.DBNameKey.String(evt.DatabaseName),
+		semconv.NetPeerNameKey.String(hostname),
+		semconv.NetPeerPortKey.Int(port),
+		semconv.NetTransportTCP,
 	}
 	if !m.cfg.CommandAttributeDisabled {
-		b, _ := bson.MarshalExtJSON(evt.Command, false, false)
-		attrs = append(attrs, DBStatement(string(b)))
+		attrs = append(attrs, semconv.DBStatementKey.String(sanitizeCommand(evt.Command)))
 	}
-
-	opts := []trace.SpanOption{
+	if collection, err := extractCollection(evt); err == nil && collection != "" {
+		spanName = collection + "."
+		attrs = append(attrs, semconv.DBMongoDBCollectionKey.String(collection))
+	}
+	spanName += evt.CommandName
+	opts := []trace.SpanStartOption{
+		trace.WithSpanKind(trace.SpanKindClient),
 		trace.WithAttributes(attrs...),
 	}
-	_, span := m.cfg.Tracer.Start(ctx, "mongodb.query", opts...)
+	_, span := m.cfg.Tracer.Start(ctx, spanName, opts...)
 	key := spanKey{
 		ConnectionID: evt.ConnectionID,
 		RequestID:    evt.RequestID,
@@ -92,20 +101,44 @@ func (m *monitor) Finished(evt *event.CommandFinishedEvent, err error) {
 	}
 
 	if err != nil {
-		span.SetAttributes(Error(true))
-		span.SetAttributes(ErrorMsg(err.Error()))
+		span.SetStatus(codes.Error, err.Error())
 	}
 
 	span.End()
 }
 
+// TODO sanitize values where possible
+// TODO limit maximum size
+func sanitizeCommand(command bson.Raw) string {
+	b, _ := bson.MarshalExtJSON(command, false, false)
+	return string(b)
+}
+
+// extractCollection extracts the collection for the given mongodb command event.
+// For CRUD operations, this is the first key/value string pair in the bson
+// document where key == "<operation>" (e.g. key == "insert").
+// For database meta-level operations, such a key may not exist.
+func extractCollection(evt *event.CommandStartedEvent) (string, error) {
+	elt, err := evt.Command.IndexErr(0)
+	if err != nil {
+		return "", err
+	}
+	if key, err := elt.KeyErr(); err == nil && key == evt.CommandName {
+		var v bson.RawValue
+		if v, err = elt.ValueErr(); err != nil || v.Type != bsontype.String {
+			return "", err
+		}
+		return v.StringValue(), nil
+	}
+	return "", fmt.Errorf("collection name not found")
+}
+
 // NewMonitor creates a new mongodb event CommandMonitor.
-func NewMonitor(serviceName string, opts ...Option) *event.CommandMonitor {
+func NewMonitor(opts ...Option) *event.CommandMonitor {
 	cfg := newConfig(opts...)
 	m := &monitor{
-		spans:       make(map[spanKey]trace.Span),
-		serviceName: serviceName,
-		cfg:         cfg,
+		spans: make(map[spanKey]trace.Span),
+		cfg:   cfg,
 	}
 	return &event.CommandMonitor{
 		Started:   m.Started,
@@ -114,14 +147,14 @@ func NewMonitor(serviceName string, opts ...Option) *event.CommandMonitor {
 	}
 }
 
-func peerInfo(evt *event.CommandStartedEvent) (hostname, port string) {
+func peerInfo(evt *event.CommandStartedEvent) (hostname string, port int) {
 	hostname = evt.ConnectionID
-	port = "27017"
+	port = 27017
 	if idx := strings.IndexByte(hostname, '['); idx >= 0 {
 		hostname = hostname[:idx]
 	}
 	if idx := strings.IndexByte(hostname, ':'); idx >= 0 {
-		port = hostname[idx+1:]
+		port = func(p int, e error) int { return p }(strconv.Atoi(hostname[idx+1:]))
 		hostname = hostname[:idx]
 	}
 	return hostname, port