+++ instrumentation/google.golang.org/grpc/otelgrpc/example/server/main.go
@@ -109,7 +109,12 @@ func (s *server) SayHelloBidiStream(stream api.HelloService_SayHelloBidiStreamSe
 }
 
 func main() {
-	config.Init()
+	tp := config.Init()
+	defer func() {
+		if err := tp.Shutdown(context.Background()); err != nil {
+			log.Printf("Error shutting down tracer provider: %v", err)
+		}
+	}()
 
 	lis, err := net.Listen("tcp", port)
 	if err != nil {